Fixed issue in earlier quoted comma fix
A previous fix changed "," to (44) to avoid issues with other comma
processing. Normally this would be a safe equivalent, but has the
unfortunate side-effect of changing LD A,"," to LD A,(44) which is
a different instruction to the expected LD A,44.

Changing the parentheses to spaces keeps the same approach, and the
token separation in expressions, but without the side-effect.