Outer parentheses now implies indirection
Parentheses are used for both indirection and to force precedence in
expressions. The acceptance depended on context, which was confusing.
LD B,(0) was not allowed due to indirection, but CP (0) was allowed
as a numeric expression.

The cases allowing expression use now treat outer parentheses as
indirection, and fail with a suitable error. In all cases the usual
work-around of 0+(expr) can stil be used.