Fixed saved size of files with DUMP < &8000
The default assembly address of &8000 was being taken as the end
address of generated code for anything assembled below &8000.
Assembling at &4000 gave a minimum size of 16K, which overwrote
the system variables when it was loaded.

The change avoids updating 'lastpage' on the first DUMP directive,
so the default location doesn't become part of the save range.