Add --intdiv option to force integer division
The '/' operator in Python 3 performs floating point division by defaut,
which is a breaking change from the integer division in Python 2. The
pyz80 expression evaluation inherits this, leading to differences in
assembly behaviour that depend on the Python version. Ouch.

To fix this we import the newer division behaviour from __future__ to
give the same float behaviour in both by default. A new --intdiv option
allows the user to force all division to be integer, matching COMET.
The alternative to this is to use the '//' operator for explicit integer
division in the code, where needed.