Fixed printing of imported symbols using -s
Symbols imported using --importfile were not added to the internal
symbolcase map, which caused an error if they were printed using -s.
The symbolcase map has been removed, and any places that used it now
build the appropriate symbol key for lookup.

The -s option now respects the --case option, and if the latter is set
only symbols that match the supplied case are printed.

Added --intdev to ReadMe.md and removed stale ReadMe.rtf file.