Fix rogue comma causing assembly failure (#20)
Adding a comma to the end of the opcode generated a type error with no
indication of what caused it, rather than an invalid opcode error message.

Use of line.split() was keeping the comma as part of the opcode, so the
eval() expression was valid but incorrect. Extracting the first word using
re.match() should be more reliable.

The previous code had a special case of recognising '(' in the opcode, then
manually splitting at that point instead. That hack is no longer needed.