Relaxed value range checking slightly (#22)
If an expression is being parsed in a known signed context no additional
range checking is performed. It doesn't make sense to truncate these values
so the caller can report the appropriate context-specific errors.

If a signed result isn't requested the input can be in the range of valid
signed or unsigned values. Values outside these ranges generate a warning,
but in all cases the result is mapped to an unsigned value.

Added more test instructions to check value truncation.