Added explicit bootable file support
Added a -B option to specify a custom bootable file, such as SAMDOS2, which is
only added to the disk image if the first assembled file isn't bootable.

This was previously possible using the -I option, but that was unconditional.
The new options fits better with the use from the VSCode extension, without
requiring setting up custom workspace settings for bootable projects.